Handle edge case for is_active in request vs response
We put all the profiling code, and also moved the existing rusage code,
behind
https://github.com/eventbrite/tikibar/blob/fff6b90e106933436cdb1c61dbcb8dde927ca17e/tikibar/middleware.py#L57-L59
in order to avoid doing that work when tikibar is not on. However, this
leads to an error the first time you make a request with tikibar enabled
after you save a change to code and django reloads it. It seems that
when process_request is hit those checks are false, but when
process_response is hit they are true, so it fails with an
attribute_error tying to read req_start_time at
https://github.com/eventbrite/tikibar/blob/fff6b90e106933436cdb1c61dbcb8dde927ca17e/tikibar/middleware.py#L93.
On subsequent page loads it works fine.

Since this is such an edge case we decided to just prevent tikibar from
loading when it occurs